backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
python-magic==0.4.27
pydantic==2.5.0
asyncio-mqtt==0.16.1